backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
open-clip-torch==2.24.0
torch==2.3.0
torchvision==0.18.0
scikit-image==0.23.2
opencv-python-headless==4.9.0.80
networkx==3.3
Pillow==10.3.0
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
reportlab==4.2.0
firebase-admin==6.5.0
